+++ apps/extension/src/stores/ui-config/manage-view-asset-token.ts
